bugfix (roundoff issues with slicing)

+++ include/cinolib/meshes/mesh_slicer.cpp
@@ -64,15 +64,15 @@ template<class Mesh>
 CINO_INLINE
 void MeshSlicer<Mesh>::update(Mesh & m, const SlicerState & s)
 {
-    float X_thresh = m.bbox().min[0] + m.bbox().delta()[0] * (s.X_thresh);
-    float Y_thresh = m.bbox().min[1] + m.bbox().delta()[1] * (s.Y_thresh);
-    float Z_thresh = m.bbox().min[2] + m.bbox().delta()[2] * (s.Z_thresh);
+    double X_thresh = m.bbox().min[0] + m.bbox().delta()[0] * (s.X_thresh);
+    double Y_thresh = m.bbox().min[1] + m.bbox().delta()[1] * (s.Y_thresh);
+    double Z_thresh = m.bbox().min[2] + m.bbox().delta()[2] * (s.Z_thresh);
 
     for(uint pid=0; pid<m.num_polys(); ++pid)
     {
-        vec3d  c = m.poly_centroid(pid);
-        float  q = m.poly_data(pid).quality;
-        int    l = m.poly_data(pid).label;
+        vec3d c = m.poly_centroid(pid);
+        float q = m.poly_data(pid).quality;
+        int   l = m.poly_data(pid).label;
 
         bool pass_X = (s.X_leq) ? (c.x() <=   X_thresh) : (c.x() >=   X_thresh);
         bool pass_Y = (s.Y_leq) ? (c.y() <=   Y_thresh) : (c.y() >=   Y_thresh);
